+++ src/pipeline.rs
@@ -2417,7 +2417,7 @@ pub fn process_packet(
         PacketAction::Rtcp(rtcp_packets) => {
             stream_store
                 .write()
-                .process_rtcp(&rtcp_packets, pp.timestamp);
+                .process_rtcp(&rtcp_packets, pp.timestamp, pp.frame);
         }
         PacketAction::Rtp {
             hdr,